requirements.txt
# 最小依赖（课程作业：CartPole 等经典控制 + 训练/画图脚本）
torch
gymnasium
numpy
pyyaml
matplotlib

# 测试（可选）
pytest

# 如需 Atari / MuJoCo / Box2D 等环境，请按需额外安装 gymnasium 对应 extra 或系统依赖。
